Major changes after the entire app wouldn't work for an excrutiating day, while I removed every file one at a time to find the culprit.
it was a react error about the element not being a valid part of the DOM.
Turned out it was because i was excluded blaze-html-templates from the
project and it then couldn't render any templates.

i cannot imagine why it worked with this file removed for the past few
days, but it did.  :-(

Partially working, but haven'tn added routes or rendering of
child components the right way.